[Sharktank] Use .to() instead of .bool(), .float(), or .int(). Remove .bool() from InferenceTensors (#2550)
Adding `.int()` and `.float()` to InferenceTensor would mean that we
clobber python's built in int and float. It breaks out type hints and
would cause strange issues if we actually try calling the functions
expecting python's implementation.

Also removing `.bool()` to make conversions more explicit.

Better to replace them all with the equivalent `.to(dtype)`.

---------

+++ sharktank/tests/models/llama4/moe_test.py
@@ -138,7 +138,9 @@ class Llama4TextMoe(torch.nn.Module):
             .scatter_(1, router_indices, router_top_value)
             .transpose(0, 1)
         )
-        router_scores = ops.sigmoid(router_scores.float()).to(hidden_states.dtype)
+        router_scores = ops.sigmoid(router_scores.to(torch.float32)).to(
+            hidden_states.dtype
+        )
 
         routed_in = hidden_states.repeat(self.num_experts, 1)
         routed_out = self.experts(routed_in).view(
